% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDM.R
\name{FDM}
\alias{FDM}
\title{Field is Mandatory or empty}
\usage{
FDM(d, mandatoryLevel = c("Mandatory", "Empty"))
}
\arguments{
\item{d}{When used in a validation rule, a bare (unquoted) name of a variable. 
Otherwise a vector of class \code{character}. Coerced to character as 
necessary.}

\item{mandatoryLevel}{character scalar indicating whether a variable
must be filled or must be empty.}
}
\description{
Field is Mandatory or empty
}
\references{
\href{../doc/20180202_maintypes.pdf}{Main types of validation rules for ESS data}: FDM
}
