% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTS.R
\name{RTS}
\alias{RTS}
\title{Check that records are present for time series}
\usage{
RTS(timevar, ftp, ltp, ...)
}
\arguments{
\item{ftp}{First time period in one of the supported notations (see Details)}

\item{ltp}{Last time period in one of the supported notations (see Details)}

\item{...}{Comma-separated list of bare (unquoted) dimensions. Time series must
be gapless from \code{ftp} to \code{ltp} for each combination of values in these
dimensions.}

\item{d}{When used in a validation rule, a bare (unquoted) name of a variable
(e.g. \code{TIME_PERIOD}). Otherwise a vector of class \code{character}.
Coerced to character as necessary.}
}
\description{
A record set is split by a set of \emph{dimensions}. For each split,
the variable indicating the time period must be both gapless and within bounds.
}
\section{Details}{


The following notations for time periods are supported:
\itemize{
  \item{\code{YYYY}: annual data, e.g. \code{"2016"}}
  \item{\code{YYYYQN}: quarterly data, e.g. \code{"2016Q1"}}
  \item{\code{YYYYMNN}: monthly data, e.g. \code{"2016M01"}, \code{"2016M10"}} 
}
}

\references{
\href{../doc/20180202_maintypes.pdf}{Main types of validation rules for ESS data}: RTS
}
