% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAD.R
\name{VAD}
\alias{VAD}
\title{Values for Aggregates are consistent with Details}
\usage{
VAD(data, field, aggregate_code, operator, tolerance, refdata)
}
\arguments{
\item{data}{a data frame when called directly. When used in a validation rule, \code{data=.} to
reference the data set passed to \code{\link[validate]{confront}}.}

\item{field}{a character scalar containing the codes to be used for this check.}

\item{aggregate_code}{a charactar scalar containing the code for the aggregate.}

\item{operator}{a charactar scalar containing a relational Operator between aggregates and the sum of detailed data}

\item{tolerance}{(acceptable margin) expressed in a absolute number}

\item{refdata}{When called directly, a data frame. When used in a validation rule, the name of the 
reference variable passed to \code{confront}.}
}
\value{
A \code{logical} vector with length the number of records. Each element of this vector contains the result of the check
}
\description{
Check that values for aggregates are consistent with the sum of values for detailed data.
}
\examples{


data(VADdat)
data(HT_AGE_GROUPSdat)

# Example in par 4.3.6 of ESTAT doc, using 'validate'
library(validate)
rule <- validator(VAD(data=., field='AGE', aggregate_code='TOTAL', operator='='
  , tolerance='0.01', ref=refdata) == TRUE)
cf <- confront(VADdat, rule, ref=list(refdata=HT_AGE_GROUPSdat))
summary(cf)
as.data.frame(cf)

# example using VAD directly

VAD(data=VADdat, field='AGE', aggregate_code='TOTAL', operator='='
  , tolerance='0.01', ref=HT_AGE_GROUPSdat)



}
\references{
\href{../doc/20180202_maintypes.pdf}{Main types of validation rules for ESS data}: VAD
}
\seealso{
Other validation-functions: \code{\link{COC}},
  \code{\link{COV}}, \code{\link{FDL}}, \code{\link{FDM}},
  \code{\link{FDT}}, \code{\link{RNR}}, \code{\link{RTS}},
  \code{\link{RWD}}, \code{\link{VCO}}, \code{\link{VIR}},
  \code{\link{VSA}}
}
\concept{validation-functions}
