% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCO.R
\name{VCO}
\alias{VCO}
\title{Values are Consistent}
\usage{
VCO(data, pivot, idvars, rule)
}
\arguments{
\item{data}{A bare (unquoted) '\code{.}' when used in a validation rule, otherwise
a data frame.}

\item{pivot}{Pivoting variable: the values \code{VAL1}, \code{VAL2},
\eqn{\ldots},\code{VALN} of this variabe become columns after pivoting.}

\item{idvars}{Identifying variables: these variables stay in rows after pivoting. All variables
that are not pivot or identifying are considered observations. A variable named \code{OBS}
will be split over columns \code{OBS.VAL1}, \code{OBS.VAL2}, \eqn{\ldots}. When used in
a validation rule, this must be a literal character. That is, its value can not be
a variable defined earlier in the script.}

\item{rule}{A bare expression in terms of the \code{OBS.VARi} variables.}
}
\description{
Pivot a table in long format around a variable, execute a validation rule
after pivoting, transform the results to the unpivoted table.
}
\examples{
data(VCOdat)

VCO(VCOdat, pivot="AGE"
 , idvars=c("TABLE","FREQ", "TIME_PERIOD","REPORTING","PARTNER","DIRECTION")
 , rule=OBS_VALUE.Y0_18/OBS_VALUE.TOTAL<0.5) 

}
\references{
\href{../doc/20180202_maintypes.pdf}{Main types of validation rules for ESS data}: VCO
}
