% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIR.R
\name{VIR}
\alias{VIR}
\title{Check that values are within a range}
\usage{
VIR(d, Min = NULL, Max = NULL)
}
\arguments{
\item{d}{When used in a validation rule, a bare (unquoted) name of a variable. 
Otherwise a vector of class \code{character}. Coerced to character as 
necessary.}

\item{Min}{smallest allowed value}

\item{Max}{largest allowed value}
}
\value{
A \code{logical} with the length of \code{d}.
}
\description{
Check that values are within a range
}
\references{
\href{../doc/20180202_maintypes.pdf}{Main types of validation rules for ESS data}: VIR
}
