% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSA.R
\name{VSA}
\alias{VSA}
\title{Check that the length of the data in the field is acceptable}
\usage{
VSA(data, tolerance = 0.01, value = "OBS_VALUE", adjust = "ADJUST",
  nsa = "N", sa = "S", idvars)
}
\arguments{
\item{data}{When used in a validation rule, a \code{.} to reference the data set passed
to \code{\link[validate]{confront}}. A data frame when called directly.}

\item{tolerance}{allowed relative difference (unadjusted data is the reference).}

\item{value}{character scalar: name of the variable with values (e.g. \code{OBS_VALUE})
column with observed values. Otherwise a numeric vector.}

\item{adjust}{character scalar: name of the variable in \code{data} indicating whether
data is seasonally adjusted or not.}

\item{nsa}{A character scalar. The code used to indicate non-seasonally adjusted data.}

\item{sa}{A character scalar. The code used to indicate seasonally adjusted data.}

\item{idvars}{character vector with names of all dimensions (identifying variables), except for 
the time period and the adjustment status.}
}
\description{
Check that the length of the data in the field is acceptable
}
\examples{
data(VSAdat)

# Using VSA directly
VSA(data= VSAdat
  , idvars = c("TABLE","FREQ","REPORTING","PARTNER","DIRECTION","AGE"))



# Using VSA with 'validate' (NOTE: data=.)
library(validate)
rules <- validator(
  VSA(data=.
    , idvars = c("TABLE","FREQ","REPORTING","PARTNER","DIRECTION","AGE")) == TRUE
)
cf <- confront(VSAdat, rules,raise="all")
summary(cf)
as.data.frame(cf)


}
\references{
\href{../doc/20180202_maintypes.pdf}{Main types of validation rules for ESS data}: VSA
}
\seealso{
Other validation-functions: \code{\link{COC}},
  \code{\link{COV}}, \code{\link{FDL}}, \code{\link{FDM}},
  \code{\link{FDT}}, \code{\link{RNR}}, \code{\link{RTS}},
  \code{\link{RWD}}, \code{\link{VAD}}, \code{\link{VCO}},
  \code{\link{VIR}}
}
\concept{validation-functions}
