% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COV.R
\name{COV}
\alias{COV}
\title{COV: Codes are Valid}
\usage{
COV(d, codelistTable = NULL, codelist = NULL)
}
\arguments{
\item{d}{When used in a validation rule, a bare (unquoted) name of a variable. 
Otherwise a vector of class \code{character}. Coerced to character as 
necessary.}

\item{codelistTable}{A character scalar: name of a csv file (US convention) 
with at least a column \code{Codes}, containing the valid codes.}

\item{codelist}{character vector with valid codes. When used in a validation
rule, this must be a literal character vector (i.e. not a variable defined
elswhere in your script).}
}
\description{
COV: Codes are Valid
}
\examples{
data(COVdat)

# Using COV with 'validate'
library(validate)
rule <- validator(COV(DIRECTION, codelist=c("IN", "OUT")) == TRUE)
cf <- confront(COVdat, rule)
summary(cf)
as.data.frame(cf)

# Using COV directly
COV(COVdat$DIRECTION, codelist=c("IN", "OUT"))

# Using 'validate' directly
rule <- validator( DIRECTION \%in\% c("IN", "OUT") )


}
\references{
\href{../doc/20180202_maintypes.pdf}{Main types of validation rules for ESS data}: COV
}
\seealso{
Other validation-functions: \code{\link{COC}},
  \code{\link{FDL}}, \code{\link{FDM}}, \code{\link{FDT}},
  \code{\link{RNR}}, \code{\link{RTS}}, \code{\link{RWD}},
  \code{\link{VAD}}, \code{\link{VCO}}, \code{\link{VIR}},
  \code{\link{VSA}}
}
\concept{validation-functions}
