% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{COVdat}
\alias{COVdat}
\alias{COCdat}
\alias{COC2dat}
\alias{COC3dat}
\alias{COC4dat}
\alias{COC5dat}
\alias{FDLdat}
\alias{FDL2dat}
\alias{FDMdat}
\alias{REPdat}
\alias{RNRdat}
\alias{RWDdat}
\alias{FDTdat}
\alias{RTSdat}
\alias{VIRdat}
\alias{RETdat}
\alias{VCOdat}
\alias{VADdat}
\alias{HT_AGE_GROUPSdat}
\title{Example data sets from the Eurostat Generic Validation Rules}
\format{csv files.}
\description{
These are the datasets from the 
\href{../doc/20180202_maintypes.pdf}{Eurostat document} where the 
generic validaiton rules are defined. They can be loaded with
\code{data(name)}, for example \code{data(FDLdat)}. The following 
datasets are available:

\itemize{
   \item{\code{COVdat}}
   \item{\code{COCdat}}
   \item{\code{COC2dat}}
   \item{\code{COC3dat}}
   \item{\code{COC4dat}}
   \item{\code{COC5dat}}
   \item{\code{FDLdat}}
   \item{\code{FDL2dat}}
   \item{\code{FDMdat}}
   \item{\code{REPdat}}
   \item{\code{RNRdat}}
   \item{\code{RWDdat}}
   \item{\code{FDLdat}}
   \item{\code{FDTdat}}
   \item{\code{RETdat}}
   \item{\code{RTSdat}}
   \item{\code{VIRdat}}
   \item{\code{VCOdat}}
   \item{\code{VADdat}}
   \item{\code{HT_AGE_GROUPSdat}}
}
}
