% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDL.R
\name{FDL}
\alias{FDL}
\title{Check that the length of the data in the field is acceptable}
\usage{
FDL(d, x = NULL, min = NULL, max = NULL, minDec = NULL,
  maxDec = NULL)
}
\arguments{
\item{d}{A character vector}

\item{x}{Number of code points required}

\item{min}{Mimimum number of code points}

\item{max}{Maximum number of code points}

\item{minDec}{minimum number of decimal positions}

\item{maxDec}{maximal number of decimal positions}
}
\description{
Check that the length of the data in the field is acceptable
}
\references{
\href{../doc/maintypes.pdf}{Main types of validation rules for ESS data}: FDL
}
