% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDL.R
\name{FDL}
\alias{FDL}
\title{Check that the length of the data in the field is acceptable}
\usage{
FDL(d, x = NULL, min = NULL, max = NULL, minDec = NULL,
  maxDec = NULL)
}
\arguments{
\item{d}{When used in a validation rule, a bare (unquoted) name of a variable. 
Otherwise a vector of class \code{character}. Coerced to character as 
necessary.}

\item{x}{Number of code points required.}

\item{min}{Mimimum number of code points}

\item{max}{Maximum number of code points}

\item{minDec}{minimum number of decimal positions}

\item{maxDec}{maximal number of decimal positions}
}
\description{
Check that the length of the data in the field is acceptable
}
\section{Details}{

The number of code points (string length in terms of human-readable characters)
may depend on current locale settings or encoding issues including those
caused by inconsistent choices of \code{UTF} normalization.
}

\references{
\href{../doc/20180202_maintypes.pdf}{Main types of validation rules for ESS data}: FDL
}
