% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDM.R
\name{FDM}
\alias{FDM}
\title{Field is Mandatory or empty}
\usage{
FDM(d, mandatoryLevel = c("Mandatory", "Empty"))
}
\arguments{
\item{d}{When used in a validation rule, a bare (unquoted) name of a variable. 
Otherwise a vector of class \code{character}. Coerced to character as 
necessary.}

\item{mandatoryLevel}{character scalar indicating whether a variable
must be filled or must be empty.}
}
\description{
Field is Mandatory or empty
}
\examples{
data(FDMdat)

# Example using 'validate' directly 
library(validate)

## OBS_VALUE is not an empty string and not NA ("Mandatory")
rules <- validator(OBS_STATUS != "" & !is.na(OBS_VALUE) )
summary( confront(FDMdat, rules) )

## OBS_VALUE is empty
rules <- validator(OBS_STATUS == "" | is.na(OBS_VALUE) )
summary( confront(FDMdat, rules) )

# Example using FDM with 'validate' 
## OBS_STATUS must be empty
rule <- validator(FDM(OBS_STATUS, mandatoryLevel="Empty") == TRUE)
cf <- confront(FDMdat, rule)
summary(cf)
as.data.frame(cf)

## OBS_STATUS is mandatory
rule <- validator(FDM(OBS_STATUS, mandatoryLevel="Mandatory") == TRUE)
cf <- confront(FDMdat, rule)
summary(cf)
as.data.frame(cf)

# Example using FDM directly
FDM(FDMdat$OBS_STATUS, mandatoryLevel="Mandatory")
FDM(FDMdat$OBS_STATUS, mandatoryLevel="Empty")


}
\references{
\href{../doc/20180202_maintypes.pdf}{Main types of validation rules for ESS data}: FDM
}
\seealso{
Other validation-functions: \code{\link{COC}},
  \code{\link{COV}}, \code{\link{FDL}}, \code{\link{FDT}},
  \code{\link{RNR}}, \code{\link{RTS}}, \code{\link{RWD}},
  \code{\link{VCO}}, \code{\link{VIR}}, \code{\link{VSA}}
}
\concept{validation-functions}
