% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTS.R
\name{RTS}
\alias{RTS}
\title{Check that records are present for time series}
\usage{
RTS(timevar, ftp, ltp, ...)
}
\arguments{
\item{timevar}{When used in a validation rule, a bare (unquoted) name of the variable
representing a time period (e.g. \code{TIME_PERIOD}). Otherwise a vector of class 
\code{character}. Coerced to character as necessary.}

\item{ftp}{First time period in one of the supported notations (see Details)}

\item{ltp}{Last time period in one of the supported notations (see Details)}

\item{...}{Comma-separated list of bare (unquoted) dimensions. Time series must
be gapless from \code{ftp} to \code{ltp} for each combination of values in these
dimensions.}
}
\description{
A record set is split by a set of \emph{dimensions}. For each split,
the variable indicating the time period must be both gapless and within bounds.
}
\section{Details}{


The following notations for time periods are supported:
\itemize{
  \item{\code{YYYY}: annual data, e.g. \code{"2016"}}
  \item{\code{YYYY-?QN}: quarterly data, e.g. \code{"2016Q1"} or \code{"2016-Q1"}}
  \item{\code{YYYYMNN}: monthly data, e.g. \code{"2016M01"}, \code{"2016M10"}} 
}
}

\examples{
# RTS examples
data(RTSdat)

# Example using RTS with 'validate'
library(validate)
rules <- validator(
 RTS(TIME_PERIOD, ftp = "2008", ltp = "2010"
   , TABLE, FREQ, REPORTING, PARTNER, DIRECTION, AGE, ADJUST) == TRUE
)
cf <- confront(RTSdat, rules)
summary(cf)
out <- as.data.frame(cf)

# Example using RTS directly
RTS(RTSdat$TIME_PERIOD, ftp = "2008", ltp = "2010"
  , RTSdat$TABLE, RTSdat$FREQ, RTSdat$REPORTING
  , RTSdat$PARTNER, RTSdat$DIRECTION, RTSdat$AGE, RTSdat$ADJUST) 

# Or, using the 'with' function from base R
with(RTSdat
   , RTS( TIME_PERIOD, ftp = "2008", ltp = "2010"
        , TABLE, FREQ, REPORTING, PARTNER, DIRECTION, AGE, ADJUST)
)

}
\references{
\href{../doc/20180202_maintypes.pdf}{Main types of validation rules for ESS data}: RTS
}
\seealso{
Other validation-functions: \code{\link{COC}},
  \code{\link{COV}}, \code{\link{FDL}}, \code{\link{FDM}},
  \code{\link{FDT}}, \code{\link{RNR}}, \code{\link{RWD}},
  \code{\link{VAD}}, \code{\link{VCO}}, \code{\link{VIR}},
  \code{\link{VSA}}
}
\concept{validation-functions}
