% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWD.R
\name{RWD}
\alias{RWD}
\alias{RWD2}
\title{Records are without duplicate ID keys.}
\usage{
RWD(...)

RWD2(...)
}
\arguments{
\item{...}{When used in a validation rule, a comma separated list of 
bare (unquoted) column names. Otherwise a comma separated list of
\code{character} vectors.}
}
\value{
For \code{RWD} a \code{logical} vector with lenght the number
of records. It is \code{FALSE} for any record that is the duplicate
of another record (with respect to the variables in the argument).

For \code{RDW2}: \code{TRUE} when there are no records that have duplicate
values for the variables in the argument, else \code{FALSE}.
}
\description{
Records are without duplicate ID keys.
}
\examples{

library(validate)
rules <- validator(RWD(height,weight)==TRUE)
cf <- confront(women, rules)
summary(cf)
as.data.frame(cf)

}
\references{
\href{../doc/20180202_maintypes.pdf}{Main types of validation rules for ESS data}: RWD
}
